{"schema":"decodex/commit/1","summary":"Modularize execution program and review boundaries","authority":"manual"}
