+++ apps/decodex/src/program_intake/issue_batch/reporting.rs
@@ -78,8 +78,7 @@ pub(in crate::program_intake) fn classify_issue(
 	}
 
 	match evaluation.lifecycle_state() {
-		ExecutionProgramNodeLifecycleState::Ready | ExecutionProgramNodeLifecycleState::Queued =>
-			IssueBatchIntakeClassification::Ready,
+		ExecutionProgramNodeLifecycleState::Ready => IssueBatchIntakeClassification::Ready,
 		ExecutionProgramNodeLifecycleState::Planned
 		| ExecutionProgramNodeLifecycleState::Mapped
 		| ExecutionProgramNodeLifecycleState::Active
